requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic==2.9.2
sentence-transformers==3.0.1
faiss-cpu
scikit-learn==1.5.1
langdetect==1.0.9
numpy==1.26.4
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
